Fix bounds bug
