--- src/components/rnd.examples.md
@@ -1,18 +1,0 @@
-Rnd:
-
-    <Rnd
-      style={{ background: '#ccc' }}
-      default={{
-        x: 100,
-        y: 100,
-        width: 320,
-        height: 200,
-      }}
-      minWidth={200}
-      minHeight={100}
-      maxWidth={800}
-      maxHeight={600}
-    >
-      Draggable
-    </Rnd>
-
